Fixed: detailPage 수정
copymain 수정
